Rendering multiple charactes
Switched to `new` function to make sure I can make separate instances of the same "class".
Updated the spritesheet to inlude two dummy characters: a boy and a grandma.
Created a simple world entity to hold everything needed for a level. Not sure if that's the way
I'll go forward, but it allows to encapsulate stuff for now. Also give a glimpse at home we can load
levels as data and then construct a Lua table out of it.