Introduce systems
After getting a better understanding of entity component system I've noticed that what I had as components
are actually systems. Components in Lua are basically entries in tables, so we don't really need special
things for them - they are part of the core language data structures.

I've also introduce state changes with transitions between state described as data. We'll have to work with this approach
a little longer to get the sense of wether it was a good implementation or not. But it made adding additional
movements to the characte straightforward, so I'm hopeful.