Add animation system
Based on frames and timers - my own naive implementation that's enough for now.
Also added another frame for a dummy boy model that turns sideways.
Introduced a concept of a flipped animation to reuse the sideways frames. And updated the rendering system to make sure it knows about
the potential animation system and can render the correct frames if they are available. A bit of refactoring but mostly
the code is still a bit messy - not a lot of consistency is happening at the moment.