Naive implementation of a dialogue bubble
Just a first experimental version. Need to figure out who's responsible for rendering the bubbles.
I will also want to have a dialogue state of the game during which the player can only control the
dialogue and not move. For now the text bubble is a good demonstration of trigger working.